Create `validator` object to give more validation flexiability
